fix: fix user invites
